requirements.txt
streamlit
keras
tensorflow
numpy
pandas
matplotlib
Pillow
setuptools
distlib